+++ server/schemas/typeDefs.js
@@ -77,7 +77,7 @@ const typeDefs = gql`
         me: User
         highScores(_id: ID!): Quiz
         countByCategory: [Category]!
-        donate(donationAmount: Float!): Donation
+        donate(donationAmount: Int!): Donation
     }
 
     type Mutation {
